Initial commit
Made a simple .py script to fetch orders data from baselinker APIand transform it to spec. Baselinker export functions where not sufficient or over the top and required a lot of additional work to accomodate for our requirements, but this just fetches the data and outputs the numbers we need. Response raw and cleaned data is saved locally and then transformed accordingly. Outputed data is just to be copy-pasted to our internal tool.